Reorganize files - split correlator from importer.
